elisp-common.eclass: add native lisp directory to load-path
There are two ways that Emacs can load elisp binaries, using
native-comp-eln-load-path, and using the normal load functions with load-path.

When using the eln-load-path method, upon loading a file, emacs will
hash the absolute directory and contents of the file, and then iterate
through eln-load-path until it finds a file in the format of
"$dirhash-$srchash-$name.eln".

Using the normal load-path and load functions, you can just append eln
to the load-suffixes, and add the path that the binaries are stored in
to load-path. With this method, upon loading a file, emacs will iterate
through the directories in load-path and search for a file in the
format of "$name.$suffix". It loops through all of the suffixes until
it finds a match.

Since we compile the elisp files in PORTAGE_TMPDIR, the hashing method
does not work correctly, so we just install the binaries into
/usr/lib(64)/emacs/ and add each packages directory to load-path
similar to what's done with normal packages in SITELISP.

To prevent loading binaries with incompatible ABIs, at install time,
we embed the ABI hash of the currently running emacs into the package's sitefile, and when the
sitefile is loaded later on by a different emacs instance, it checks
if the ABI hash matches the embedded value before
appending load-path.